+++ bittytax/conv/parsers/cryptopia.py
@@ -9,6 +9,8 @@ from ..exceptions import UnexpectedTypeError
 
 WALLET = "Cryptopia"
 
+PRECISION = Decimal('0.00000000')
+
 def parse_cryptopia_deposits(data_row, _parser, _filename):
     in_row = data_row.in_row
     data_row.timestamp = DataParser.parse_timestamp(in_row[7])
@@ -36,21 +38,23 @@ def parse_cryptopia_trades(data_row, parser, _filename):
     data_row.timestamp = DataParser.parse_timestamp(in_row[7])
 
     if in_row[2] == "Buy":
+        sell_quantity = Decimal(in_row[3]) * Decimal(in_row[4])
+
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
                                                  data_row.timestamp,
                                                  buy_quantity=in_row[4],
                                                  buy_asset=in_row[1].split('/')[0],
-                                                 sell_quantity=Decimal(in_row[3]) * \
-                                                               Decimal(in_row[4]),
+                                                 sell_quantity=sell_quantity.quantize(PRECISION),
                                                  sell_asset=in_row[1].split('/')[1],
                                                  fee_quantity=in_row[6],
                                                  fee_asset=in_row[1].split('/')[0],
                                                  wallet=WALLET)
     elif in_row[2] == "Sell":
+        buy_quantity = Decimal(in_row[3]) * Decimal(in_row[4])
+
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_TRADE,
                                                  data_row.timestamp,
-                                                 buy_quantity=Decimal(in_row[3]) * \
-                                                              Decimal(in_row[4]),
+                                                 buy_quantity=buy_quantity.quantize(PRECISION),
                                                  buy_asset=in_row[1].split('/')[1],
                                                  sell_quantity=in_row[4],
                                                  sell_asset=in_row[1].split('/')[0],