+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Cell conversion of imported Excel data safer for python 2.
 - Circle parser: filter out other currency symbols '£€$'.
 - Cryptsy parser: sell/buy quantities already had fee included.
+- Cryptopia parser: calculations rounded to 8 decimal places.
 ### Added
 - Conversion tool: added parser for CoinTracking.info accounting data.
 - Conversion tool: added parser for Gravity (Bitstocks) exchange.